* Added Blessing of Salvation to the list of possible Normal Blessings to be used as an Alternate Blessing or in a Raid to wipe off Greater Blessing of Salvation from tanks.
* Removed the IsInRaid() disablement check from the Raid tab in the PallyPower Options panel.

* Removed the additional raid warnings from the Blessings Report that would appear if there are 5 or more Paladins in a raid.

SPECIAL NOTE: Tanks should NOT be manually switching off Blessing of Salvation due to recent issues with the LibThreatClassic2 library not re-calculating Threat when a tank manually clicks off Blessing of Salvation. I'm sure it'll be fixed but I'm also sure that a similar bug might appear at a later date. Tanks and Paladins should begin getting used to the idea of allowing/requiring Paladins to remove their Salvation.